Moving get_split_lists to FASTA.py
It just makes more sense to have all functionalities related to the fasta files in one place
